refactor: inline console logger and simplify cache
