--- src/dns-shim.d.ts
@@ -1,5 +1,0 @@
-declare module 'dns' {
-  export const promises: any;
-}
-
-declare const process: any;
